ビルド警告の修正

+++ recfriior5/Hdp.hpp
@@ -78,7 +78,7 @@ private:
     uint8_t r5;
 
     void sendTuner(uint8_t *data, int len);
-    void send_ctl(uint8_t *data, int fd, int req, int val, int idx, int len);
+    void send_ctl(uint8_t *data, int fd, uint8_t req, uint16_t val, uint16_t idx, uint16_t len);
 
     bool vendorRequest ( uint16_t wLength, uint8_t* data, uint8_t bRequest, uint8_t v0 = 0, uint8_t v1 = 0, uint8_t v2 = 0, uint8_t v3 = 0 );
